added sqlite3 ebookstore task and database files
This project demonstrates how sqlite can be useful to a bookstore clerk. 
It allows a clerk to add new books, update book information, search and delete books from a sql database.
It is structured to output a menu as well as format the requested book data in a readable manner to the clerk.